+++ src/Data/Interval/Borel.hs
@@ -68,40 +68,40 @@ instance (Ord x) => Monoid (Borel x) where
   mempty :: (Ord x) => Borel x
   mempty = Borel mempty
 
-instance (Ord x, Lattice x) => Lattice (Borel x) where
-  (\/) :: (Ord x, Lattice x) => Borel x -> Borel x -> Borel x
+instance (Ord x) => Lattice (Borel x) where
+  (\/) :: (Ord x) => Borel x -> Borel x -> Borel x
   (\/) = union
 
-  (/\) :: (Ord x, Lattice x) => Borel x -> Borel x -> Borel x
+  (/\) :: (Ord x) => Borel x -> Borel x -> Borel x
   (/\) = intersection
 
-instance (Ord x, Lattice x) => BoundedMeetSemiLattice (Borel x) where
-  top :: (Ord x, Lattice x) => Borel x
+instance (Ord x) => BoundedMeetSemiLattice (Borel x) where
+  top :: (Ord x) => Borel x
   top = whole
 
-instance (Ord x, Lattice x) => BoundedJoinSemiLattice (Borel x) where
-  bottom :: (Ord x, Lattice x) => Borel x
+instance (Ord x) => BoundedJoinSemiLattice (Borel x) where
+  bottom :: (Ord x) => Borel x
   bottom = mempty
 
-instance (Ord x, Lattice x) => Heyting (Borel x) where
-  (==>) :: (Ord x, Lattice x) => Borel x -> Borel x -> Borel x
+instance (Ord x) => Heyting (Borel x) where
+  (==>) :: (Ord x) => Borel x -> Borel x -> Borel x
   x ==> y = complement x \/ y
 
-instance (Ord x, Lattice x) => Semiring (Borel x) where
-  plus :: (Ord x, Lattice x) => Borel x -> Borel x -> Borel x
+instance (Ord x) => Semiring (Borel x) where
+  plus :: (Ord x) => Borel x -> Borel x -> Borel x
   plus = symmetricDifference
 
-  times :: (Ord x, Lattice x) => Borel x -> Borel x -> Borel x
+  times :: (Ord x) => Borel x -> Borel x -> Borel x
   times = intersection
 
-  zero :: (Ord x, Lattice x) => Borel x
+  zero :: (Ord x) => Borel x
   zero = mempty
 
-  one :: (Ord x, Lattice x) => Borel x
+  one :: (Ord x) => Borel x
   one = whole
 
-instance (Ord x, Lattice x) => Ring (Borel x) where
-  negate :: (Ord x, Lattice x) => Borel x -> Borel x
+instance (Ord x) => Ring (Borel x) where
+  negate :: (Ord x) => Borel x -> Borel x
   negate = complement
 
 -- | Consider the 'Borel' set identified by a list of 'Interval's.